New types update
Changed the way nodes are defined in both pydantic and typescript, so that their structures match.

Conforming to this removed the need for functions that reshape the json/pydantic/TS interface structure when passing data between frontend and backend.

Changing the way typescript handles these also allowed the typescript build to succeed.