thumbnail functionality in the inspector and backend
When working with images, it's probably not the best to be sending those images in array form to and from the frontend. Thumbnails can be automatically generated, and expanding the thumbnail into a modal will trigger a fetch of the full image and display it. This can apply to other potentially large data types as well, Arrays, gcode file, etc.